% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{all_cn_chr}
\alias{all_cn_chr}
\title{Regex-Syntax for subsetting all Chinese characters.}
\format{
## `all_cn_chr`
a chr string of regex syntax to match all Chinese characters.
}
\usage{
all_cn_chr
}
\description{
Regex-Syntax for subsetting all Chinese characters.
}
\keyword{datasets}
