% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cncrs}
\alias{cncrs}
\title{PROJ4 string of China projection coordinate system.}
\format{
## `cncrs`
a string of PROJ4 Albert projection for China domain.
}
\usage{
cncrs
}
\description{
PROJ4 string of China projection coordinate system.
}
\keyword{datasets}
