% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cnmap_9dx}
\alias{cnmap_9dx}
\title{Nine dotted lines for China south sea boundary.}
\format{
a sf line object,
}
\usage{
cnmap_9dx
}
\description{
Nine dotted lines for China south sea boundary.
}
\keyword{datasets}
