% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cnmap_counties}
\alias{cnmap_counties}
\title{China map in county scale, imported and simplified form `mapchina` package.}
\format{
a sf polygon object.
}
\usage{
cnmap_counties
}
\description{
China map in county scale, imported and simplified form `mapchina` package.
}
\keyword{datasets}
