% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cnmap_islands}
\alias{cnmap_islands}
\title{islands of China in far sea which are not included in county-scale map.}
\format{
a sf polygon object.
}
\usage{
cnmap_islands
}
\description{
islands of China in far sea which are not included in county-scale map.
}
\keyword{datasets}
