% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{cnmap_simplified}
\alias{cnmap_simplified}
\title{Simplified China boundary for mainland and Taiwan.}
\format{
a sf polygon object.
}
\usage{
cnmap_simplified
}
\description{
Simplified China boundary for mainland and Taiwan.
}
\keyword{datasets}
