% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{convert_amount}
\alias{convert_amount}
\title{Convert value between common Chinese quantifiers}
\usage{
convert_amount(value, old.unit, new.unit)
}
\arguments{
\item{value}{num, the value to be converted}

\item{old.unit}{chr, the old unit}

\item{new.unit}{chr, the new unit}
}
\value{
a number of converted value in the new unit
}
\description{
Convert value between common Chinese quantifiers
}
\examples{
convert_amount(133, "兆瓦", "万千瓦")
}
