% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{convert_coord}
\alias{convert_coord}
\title{Convert a geo-coding of between common Chinese coordinate systems}
\usage{
convert_coord(lng, lat, from, to)
}
\arguments{
\item{lng}{num, longitude}

\item{lat}{num, latitude}

\item{from}{chr, source coord system, supports "BD09", "GCJ02" and "WGS84"}

\item{to}{chr, target coord system, supports "BD09", "GCJ02" and "WGS84"}
}
\value{
2 number of geocoding in the converted coordinate system
}
\description{
Convert a geo-coding of between common Chinese coordinate systems
}
\examples{
convert_coord(133, 47, from = "GCJ02", to = "BD09")
}
