% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{DaliyMeanConc}
\alias{DaliyMeanConc}
\title{Calculate daily mean concentration of 6 pollutants}
\usage{
DaliyMeanConc(Pollu, Conc)
}
\arguments{
\item{Pollu}{name of pollutants now supports `SO2`, `NO2`, `CO`, `PM2.5`, `PM10` and `O3`}

\item{Conc}{a 24-long vector of hourly average concentrations}
}
\value{
a number of daily mean concentration of 6 pollutants in which `O3` is calculated by maximum 8-hour average
}
\description{
Calculate daily mean concentration of 6 pollutants
}
\examples{
C <- rnorm(24, 35, 5)
DaliyMeanConc("PM2.5", C)
}
