% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evan_ggtheme.R
\name{Evantheme}
\alias{Evantheme}
\title{My ggplot2 theme for scientific plotting use}
\usage{
Evantheme()
}
\value{
a ggplot theme object
}
\description{
My ggplot2 theme for scientific plotting use
}
\examples{
ggplot(mtcars, aes(x = mpg, y = cyl)) + geom_point() + stat_smooth() + Evantheme()
}
