% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{ffmpeg_video.compress}
\alias{ffmpeg_video.compress}
\title{Call ffmpeg in shell to compress video}
\usage{
ffmpeg_video.compress(
  video,
  resolution = 720,
  bit_rate = 2000,
  encoder = "hevc_nvenc",
  extra = ""
)
}
\arguments{
\item{video}{video path, full path.}

\item{resolution}{portrait video picture size, i.e., 480, 720, 1080, 1440, 2160.}

\item{bit_rate}{bit rate, unit:k.}

\item{encoder}{video encoder, by default `hevc_nvenc` for nvidia GPU. alternative `hevc_amf` for AMD GPU. `libx264` for CPU encoding only.}

\item{extra}{extra option pass to ffmpeg.}
}
\value{
status for compile.
}
\description{
Call ffmpeg in shell to compress video
}
\examples{
ffmpeg_video.compress("path/to/your/video")
}
