% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{force_path_validate}
\alias{force_path_validate}
\title{Check if the path exist, if not, create every dir in the path.}
\usage{
force_path_validate(path)
}
\arguments{
\item{path}{A string of file path that you want to use for save a file.
Note that it must direct to a file not a folder.}
}
\value{
the checked `path` same as input
}
\description{
Check if the path exist, if not, create every dir in the path.
}
\examples{
force_path_validate("~/THE/FILE/PATH/THAT/YOU/WANT/TO/TEST.TXT")
}
