% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{get_ext}
\alias{get_ext}
\title{Get file extension.}
\usage{
get_ext(fname)
}
\arguments{
\item{fname}{file name/path.}
}
\value{
file ext, or 0 length character if no extension detected.
}
\description{
Get file extension.
}
\examples{
writeLines("","test.R")
get_ext("./test.R")
unlink("test.R")
# run in Rstudio script panel:
\dontrun{
 rstudioapi::getActiveDocumentContext()$path |> get_fname(keep.ext = TRUE)
}
}
