% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{get_fileext}
\alias{get_fileext}
\title{Get file extension.}
\usage{
get_fileext(fname)
}
\arguments{
\item{fname}{file name/path.}
}
\value{
file ext, or 0 length character if no extension detected.
}
\description{
Get file extension.
}
\examples{
# run in Rstudio script panel:
\dontrun{
  writeLines("","test.R")
  get_fileext("./test.R")
  unlink("test.R")
  rstudioapi::getActiveDocumentContext()$path |> get_fname(keep.ext = TRUE)
}
}
