% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{get_fname}
\alias{get_fname}
\title{Get filename from a path,}
\usage{
get_fname(path, ftype = "R", no_abbr = T)
}
\arguments{
\item{path}{Path to a file (Attention: NOT A DIR). String of file path.}

\item{ftype}{Character, string of file extension, default `R` for Rscript file}

\item{no_abbr}{Boolean, control option for out put file name: should file extension be included, by default TRUE, represents not including.}
}
\value{
a string of file name without file extension abbr.
}
\description{
Get filename from a path,
}
\examples{
writeLines("","test.R")
get_fname("./test.R")
unlink("test.R")
}
