% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{get_grayscale}
\alias{get_grayscale}
\title{Get gray scale of color(s), useful when deciding color of text above a background color.}
\usage{
get_grayscale(color)
}
\arguments{
\item{color}{vector of character, hex RGB color or default color space.}
}
\value{
grayscale from 0(pure black) to 100 (pure bright)
}
\description{
Get gray scale of color(s), useful when deciding color of text above a background color.
}
\examples{
get_grayscale("red")

get_grayscale(c("red", "yellow", "green"))

\donotrun{
  library(tidyverse)
  df1 <- matrix(colors()[1:120], ncol = 6) \%>\%  as.data.frame() \%>\% rownames_to_column("Row") \%>\%
    pivot_longer(cols = -1, names_to = "Col", names_prefix = "V") \%>\%
    mutate(across(Row:Col, as.numeric))
  ggplot(df1, aes(x = Col, y = Row)) +
    geom_raster(aes(fill = value)) +
    geom_text(aes(label = value, color = if_else(get_grayscale(value) > 60, "black", "white"))) +
    scale_fill_identity() +
    scale_color_identity() +
    theme_void()
}
}
