% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQI.R
\name{IAQI_Daily}
\alias{IAQI_Daily}
\title{Calculate daily IAQI for given pollutant}
\usage{
IAQI_Daily(Pollu, Conc)
}
\arguments{
\item{Pollu}{name of pollutants now supports `SO2`, `NO2`, `CO`, `O3`, `PM2.5` and `PM10`.}

\item{Conc}{a number of daily mean concentration of the pollutant.}
}
\value{
the daily IAQI index score of the give pollutant.
}
\description{
Calculate daily IAQI for given pollutant
}
\examples{
IAQI_Daily("O3", 142)

}
