% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utillities.R
\name{llm_join}
\alias{llm_join}
\title{Title JOIN by LLMs
Generate prompt to do fuzzy join with LLMs. By far(2025/04/10), DeepSeek R1 showed the best result,
other LLMs might fabricate non-existing data in the result.}
\usage{
llm_join(tb1, tb2, key1, key2)
}
\arguments{
\item{tb1}{`data.frame`, left data.frame}

\item{tb2}{`data.frame`, right data.frame}

\item{key1}{key column in the left data.frame}

\item{key2}{key column in the right data.frame}
}
\value{
prompt that used to tell the LLMs
}
\description{
Title JOIN by LLMs
Generate prompt to do fuzzy join with LLMs. By far(2025/04/10), DeepSeek R1 showed the best result,
other LLMs might fabricate non-existing data in the result.
}
\examples{
#' # run in Rstudio script panel:
\dontrun{
  library(ggplot2)
  library(maps)
  world <- map_data("world")
  df <- LifeCycleSavings
  df["country"] = rownames(df)
  llm_join(world, df, "region","country")
}
}
