% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{mk_path}
\alias{mk_path}
\title{Check if the path exist, if not, create every dir in the path.}
\usage{
mk_path(path)
}
\arguments{
\item{path}{A string of file path that you want to use for save a file.
Note that it must direct to a file not a folder.}
}
\value{
the checked `path` same as input
}
\description{
Check if the path exist, if not, create every dir in the path.
}
\examples{
mk_path("~/THE/FILE/PATH/THAT/YOU/WANT/TO/TEST.TXT")
}
