% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utillities.R
\name{mkdir}
\alias{mkdir}
\title{Check if the path exist, if not, create every dir in the path.}
\usage{
mkdir(path)
}
\arguments{
\item{path}{A string of file path that you want to use for save a file.}
}
\value{
the checked `path`
}
\description{
Check if the path exist, if not, create every dir in the path.
}
\examples{
mkdir("~/THE/FILE/PATH/THAT/YOU/WANT/TO/TEST.TXT")
}
