% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{simple_date}
\alias{simple_date}
\title{Get a simple date for document naming}
\usage{
simple_date(x = NULL)
}
\arguments{
\item{x}{`NULL` or a `Date`, if NULL, returns the present date.}
}
\value{
A 6-number long string, first 2 is the year; middle 2 is the month; last 2 is the day.
}
\description{
Get a simple date for document naming
}
\examples{
simple_date()
simple_date(as.Date("2023-09-09"))
}
