% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{split_by}
\alias{split_by}
\title{Split dataframe into list of dataframes based on selection.}
\usage{
split_by(.data, ..., keep = F)
}
\arguments{
\item{.data}{A `data.frame`, either a `tibble` or a `data.frame` is okay.}

\item{...}{`<tidy-select>` syntax for picking out column(s) to split.}

\item{keep}{A Boolean value to decide whether the column should kept in the splited data.frame.}
}
\value{
A named-list consists of pieces of `data.frame`s.
}
\description{
Split dataframe into list of dataframes based on selection.
}
\examples{
CO2 |> split_by(Plant,Type)
}
