% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utillities.R
\name{video.compress}
\alias{video.compress}
\title{Call ffmpeg in shell to compress video}
\usage{
video.compress(
  video,
  res = NULL,
  bv = NULL,
  force_format = NULL,
  cv = "hevc_nvenc",
  extra = ""
)
}
\arguments{
\item{video}{video path, full path.}

\item{res}{resolution, numeric. especially portrait video picture size, i.e., 480, 720, 1080, 1440, 2160.
By default NULL for not changing bit rate.}

\item{bv}{video bit rate, unit: kb. By default NULL for not changing bit rate.}

\item{force_format}{Output file format, chr. By default NULL for keeping file format as input file.}

\item{cv}{video encoder, by default `hevc_nvenc` for nvidia GPU hwa. alternative `hevc_amf` for AMD GPU hwa.
`libx264` for CPU encoding only.}

\item{extra}{extra option pass to ffmpeg.}
}
\value{
status for compile.
}
\description{
Call ffmpeg in shell to compress video
}
\examples{
ffmpeg_video.compress("path/to/your/video")
}
