% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{get_fname}
\alias{get_fname}
\title{Get file name from a path, without file ext or with modified file ext.}
\usage{
get_fname(path, new.ext = "")
}
\arguments{
\item{path}{Path to a file (Attention: NOT A DIR). String of file path.}

\item{new.ext}{String, for the designated file extension.
By default `""` represent not changing anything. If any string was provided, the returned
file name will be attached with the new file extension.}
}
\value{
String, file name without file ext, or with modified file ext.
}
\description{
Get file name from a path, without file ext or with modified file ext.
}
\examples{
writeLines("","test.R")
get_fname("./test.R")
unlink("test.R")
# run in Rstudio script panel:
\dontrun{
 rstudioapi::getActiveDocumentContext()$path |> get_fname(keep.ext = TRUE)
}
}
