% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get.R
\name{get_fname}
\alias{get_fname}
\title{Get filename from a path,}
\usage{
get_fname(path, keep.ext = F)
}
\arguments{
\item{path}{Path to a file (Attention: NOT A DIR). String of file path.}

\item{keep.ext}{Boolean, should the file extension be included. By default not include.}
}
\value{
a string of file name without file extension abbr.
}
\description{
Get filename from a path,
}
\examples{
writeLines("","test.R")
get_fname("./test.R")
unlink("test.R")
# run in Rstudio script panel:
\dontrun{
 rstudioapi::getActiveDocumentContext()$path |> get_fname(keep.ext = TRUE)
}
}
