% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{larger_bbox}
\alias{larger_bbox}
\title{get a more suitable bounding box of a sf/sfc object according to the precising for raster-making}
\usage{
larger_bbox(x, precise = 0.25)
}
\arguments{
\item{x}{the sf/sfc object}

\item{precise}{the precise/size of the ideal raster}
}
\value{
a sf-bbox object
}
\description{
get a more suitable bounding box of a sf/sfc object according to the precising for raster-making
}
\examples{
sf::st_bbox(mapchina::china)
larger_bbox(mapchina::china, precise = 0.25)
}
