% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{simple_date}
\alias{simple_date}
\title{get a simple date for document naming}
\usage{
simple_date()
}
\value{
6-charactor long numbers first 2 is the year; middle 2 is the month; last 2 is the date.
}
\description{
get a simple date for document naming
}
\examples{
simple_date()
}
