% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate.R
\name{simple_date}
\alias{simple_date}
\title{get a simple date for document naming}
\usage{
simple_date(x = NULL)
}
\arguments{
\item{x}{`NULL` or a `Date` format num, if NULL, returns the present date.}
}
\value{
6-charactor long numbers first 2 is the year; middle 2 is the month; last 2 is the date.
}
\description{
get a simple date for document naming
}
\examples{
simple_date()
simple_date(as.Date("2023-09-09"))
}
